requirements.txt
selene >= 2.0.0rc8
pytest >= 8.0.2
allure-pytest >= 2.13.2